feat: 예약 삭제 시 로그인 여부 확인 기능 구현
- 예약 테스트 시 로그인 과정 추가
- 전체 코드 리포맷팅

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,16 +1,15 @@
 package nextstep.reservation;
 
+import java.util.List;
 import java.util.Objects;
-import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class ReservationService {
@@ -57,10 +56,13 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+        if (!Objects.equals(reservation.getMemberId(), memberId)) {
+            throw new AuthorizationException();
         }
 
         reservationDao.deleteById(id);
